Fixes for 2 snippet samples in Deconstructing tuples and other types (#28)
*  Missing out from second Deconstruct method

The second Deconstruct method was missing an out before the last parameter decimal income.

* Moved snippet 1 comment

In Deconstructing user-defined types the sentence preceding the first snippet says "For example, the following Deconstruct method..." it was showing the constructor and not the Deconstruct method in the code snippet.  Credit goes to vishaalkal for pointing it out in the comments.