Clarify DateTime implicit conversion example (#850)
* Clarify DateTime implicit conversion example

It took some time for me to understand why the last example has -08:00 as offset. Since the example seemed to be about the addition of a TimeSpan the change in offset was quite confusing. 

Therefore, I split the example into two parts, one only focusing on the addition of a TimeSpan and the second on the adaption of the daylight time specific time zone. 

I also fixed a typo in the year given in line 84.

* Clarify DateTime implicit conversion VB example

Adaption of changes made to the C# example in commit e1833f10f89ccbf24cce277a52fcd3601e3173c8 to the Visual Basic sample base.
This change was made based on the feedback received to [PR #850](https://github.com/dotnet/samples/pull/850#issuecomment-487091258)