Fix exception when sample is built for .NET Core debug. (#858)
If you try this sample in a .NET Core project, and build as debug, it will throw an exception on line 27. This is because, when built for debug, the value is not garbage collected, so `wr2.Target` is not `null` and we hit the `else if`. In that line, `mc2` *is* `null` (because it is set to `null` on line 19) which throws an exception trying to use it as a key to the CWT.