make snippets for samples (#960)
Contributes to dotnet/dotnet-api-docs#2539

For `String.Split`, almost all the samples had been changed to interactive. However, most were full programs. These were turned into snippets.

In one case, a static function was changed to a local function.

In two other cases, The `System.Text.RegularExpressions` namespace was added to variable declarations.

The dotnet-api-docs content does not need to be changed, so this may need to be verified after merging by forcing a new build in the portal.