Use EnableDynamicLoadings for all plugins
In .NET 6 SDk doesn't produce `.runtimeconfig.dev.json` by default and as such plugins with NuGet dependencies won't run with just `dotnet build` as they don't copy their dependencies to output.

Adding `EnableDynamicLoading=true` to all plugin projects to fix this. Technically it's not necessary, and just `CopyLocalLockFileAssemblies=true` would be enough, but I think it's just really weird. Also it's much easier to explain: "I want to load this project dynamically at runtime, so I'm going to mark it with `EnableDynamicLoading=true`".

Removed the `XcopyablePlugin` since it's not useful anymore, all of the plugins are now "xcopyable".